+++ BitFaster.Caching/Lru/Builder/ScopedLruBuilder.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BitFaster.Caching.Lru.Builder
+{
+    public class ScopedLruBuilder<K, V, W> : LruBuilderBase<K, V, ScopedLruBuilder<K, V, W>, IScopedCache<K, V>> where V : IDisposable where W : IScoped<V>
+    {
+        private readonly ConcurrentLruBuilder<K, W> inner;
+
+        internal ScopedLruBuilder(ConcurrentLruBuilder<K, W> inner)
+            : base(inner.info)
+        {
+            this.inner = inner;
+        }
+
+        public override IScopedCache<K, V> Build()
+        {
+            // this is a legal type conversion due to the generic constraint on W
+            var scopedInnerCache = inner.Build() as ICache<K, Scoped<V>>;
+
+            return new ScopedCache<K, V>(scopedInnerCache);
+        }
+    }
+}
