+++ BitFaster.Caching/Scoped.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading;
+using BitFaster.Caching.Lru;
 
 namespace BitFaster.Caching
 {
